+++ modules/rest/v1/settings.php
