+++ package.json
@@ -6,7 +6,7 @@
   "type": "module",
   "scripts": {
     "test": "echo 'Error: no test specified' && exit 1",
-    "build:css": "tailwindcss -i admin/settings/Views/src/input.css -o admin/assets/css/build/main.css",
+    "build:css": "tailwindcss -i admin/settings/views/src/input.css -o admin/assets/css/build/main.css",
     "build": "webpack --config webpack.config.js --mode=production",
     "start": "wp-scripts start modules/wizard/src/setup.js --output-path admin/assets/frontend/setup",
     "start:js": "webpack --config webpack.config.js --mode=development --watch",
